+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+### 0.4
 - Add option to automatically lock the levels of newly plopped buildings.
 - Add workaround for compatibility with Land Value Overhaul mod.